get full graph with create graph functionality

+++ src/main/java/org/akshay/campusnavigator/graph/AdjacencyListNode.java
@@ -1,29 +1,29 @@
 package org.akshay.campusnavigator.graph;
 
 public class AdjacencyListNode {
-    public int sourceNode;
-    public int destinationNode;
-    public int distance;
+    public Long sourceNode;
+    public Long destinationNode;
+    public Double distance;
     AdjacencyListNode next;
 
-    public AdjacencyListNode(int sourceNode, int destinationNode, int distance, AdjacencyListNode next) {
+    public AdjacencyListNode(Long sourceNode, Long destinationNode, Double distance, AdjacencyListNode next) {
         this.sourceNode = sourceNode;
         this.destinationNode = destinationNode;
         this.distance = distance;
         this.next = next;
     }
 
-    public AdjacencyListNode(int sourceNode, int destinationNode, int distance) {
+    public AdjacencyListNode(Long sourceNode, Long destinationNode, Double distance) {
         this.sourceNode = sourceNode;
         this.destinationNode = destinationNode;
         this.distance = distance;
         this.next = null;
     }
 
-    public AdjacencyListNode(int sourceNode, int destinationNode) {
+    public AdjacencyListNode(Long sourceNode, Long destinationNode) {
         this.sourceNode = sourceNode;
         this.destinationNode = destinationNode;
-        this.distance = 0; // For nodes inside building
+        this.distance = 0.0; // For nodes inside building
         this.next = null;
     }
